// apps/web/src/shared/api/generated/foundation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Tomotabi M0 foundation
 * Compatibility probe only. Not a product feature.
 * OpenAPI spec version: 0.0.0
 */
import { httpClient } from '../http-client';
export interface ProbeView {
  /** @minimum 0 */
  count: number;
}

export type GetHealth200 = {
  status: 'ok';
};

export type getHealthResponse200 = {
  data: GetHealth200
  status: 200
}

export type getHealthResponseSuccess = (getHealthResponse200) & {
  headers: Headers;
};
;

export type getHealthResponse = (getHealthResponseSuccess)

export const getGetHealthUrl = () => {




  return `/api/health`
}

export const getHealth = async ( options?: Parameters<typeof httpClient>[1]): Promise<getHealthResponse> => {

  return httpClient<getHealthResponse>(getGetHealthUrl(),
  {
    ...options,
    method: 'GET'


  }
);}



export type getProbeResponse200 = {
  data: ProbeView
  status: 200
}

export type getProbeResponseSuccess = (getProbeResponse200) & {
  headers: Headers;
};
;

export type getProbeResponse = (getProbeResponseSuccess)

export const getGetProbeUrl = () => {




  return `/api/foundation/probes`
}

export const getProbe = async ( options?: Parameters<typeof httpClient>[1]): Promise<getProbeResponse> => {

  return httpClient<getProbeResponse>(getGetProbeUrl(),
  {
    ...options,
    method: 'GET'


  }
);}



export type incrementProbeResponse200 = {
  data: ProbeView
  status: 200
}

export type incrementProbeResponseSuccess = (incrementProbeResponse200) & {
  headers: Headers;
};
;

export type incrementProbeResponse = (incrementProbeResponseSuccess)

export const getIncrementProbeUrl = () => {




  return `/api/foundation/probes/increment`
}

export const incrementProbe = async ( options?: Parameters<typeof httpClient>[1]): Promise<incrementProbeResponse> => {

  return httpClient<incrementProbeResponse>(getIncrementProbeUrl(),
  {
    ...options,
    method: 'POST'


  }
);}